Adding RubyPython::Tuple.
To simplify cases where actual Python tuples are required, adding a
class (a subclass of ::Array) to allow for explicit conversion to tuple
objects.

The semantics are best handled with the class method
RubyPython::Tuple.tuple(), which makes a #dup of the Array parameter.
This matches the result of the tuple() method in Python.